refactor(ui): move terminal dialog form to a separated component
