+++ ui/src/interfaces/ITerminal.ts
@@ -1,5 +1,7 @@
 export interface IConnectToTerminal {
+  username: string;
   password?: string;
+  privateKey?: string;
   signature?: string;
   fingerprint?: string;
 }
